challenge_1/challenge_1/cloud_entrypoint/requirements.txt
tensorflow
tensorflow-cloud
google-cloud
